fix(patina): keep over-limit files from growing
`tests/tooling/source-file-lengths` failed on this PR because two already
over-limit files grew: `crates/vize_patina/src/linter/tests/jsx.rs` (422 ->
475) and `crates/vize_patina/src/rule.rs` (439 -> 440).

Move the three new JSX fallback tests into a sibling `jsx_fallback`
module so `jsx.rs` returns to its base length, and factor the default
preset's security-rule registrations into a `register_security` helper
in `rules/vue` so `rule.rs` shrinks below its base.
