+++ crates/vize_patina/src/linter/tests.rs
@@ -5,5 +5,6 @@ use vize_carton::{Allocator, ToCompactString};
 mod basic;
 mod directives;
 mod jsx;
+mod jsx_fallback;
 mod no_top_level_ref;
 mod sfc;
